refactor: AdjPostNav -> PostFooter rename
