+++ app/[locale]/community/news/[id]/page.tsx
@@ -5,10 +5,11 @@ import { StraightNode } from '@/components/common/Nodes';
 import Tags from '@/components/common/Tags';
 import HTMLViewer from '@/components/editor/HTMLViewer';
 import PageLayout, { PAGE_PADDING_BOTTOM_PX } from '@/components/layout/pageLayout/PageLayout';
-import AdjPostNav from '@/components/post/AdjPostNav';
+import PostFooter from '@/components/post/PostFooter';
 
 import { PostSearchQueryParams } from '@/types/post';
 
+import { formatNewsPostDateStr } from '@/utils/date';
 import { getPath } from '@/utils/page';
 import { news } from '@/utils/segmentNode';
 
@@ -22,11 +23,6 @@ const newsPath = getPath(news);
 export default async function NewsPostPage({ params, searchParams }: NewsPostPageProps) {
   const news = await getNewsPostDetail(parseInt(params.id), searchParams);
 
-  const date = new Date(news.date);
-  const dateStr = `${date.getFullYear()}년 ${
-    date.getMonth() + 1
-  }월 ${date.getDate()}일 ${date.toLocaleString('ko-KR', { weekday: 'long' })}`;
-
   return (
     <PageLayout titleType="big" bodyStyle={{ padding: 0 }}>
       <h2 className="px-5 py-9 text-[1.25rem] font-semibold sm:pl-[100px] sm:pr-[340px]">
@@ -35,16 +31,16 @@ export default async function NewsPostPage({ params, searchParams }: NewsPostPag
 
       <div
         className="bg-neutral-50 px-5 pt-9 sm:pl-[100px] sm:pr-[340px]"
-        style={{
-          paddingBottom: PAGE_PADDING_BOTTOM_PX,
-        }}
+        style={{ paddingBottom: PAGE_PADDING_BOTTOM_PX }}
       >
         <Attachments files={news.attachments} />
         <HTMLViewer htmlContent={news.description} className="mb-10" />
-        <time className="mb-3 mt-12 block text-end text-sm font-bold">{dateStr}</time>
+        <time className="mb-3 mt-12 block text-end text-sm font-bold">
+          {formatNewsPostDateStr(news.date)}
+        </time>
         <StraightNode />
         <Tags tags={news.tags} margin="mt-3 ml-6" searchPath={newsPath} />
-        <AdjPostNav post={news} postType="notice" id={params.id} margin="mt-12" />
+        <PostFooter post={news} postType="news" id={params.id} margin="mt-12" />
       </div>
     </PageLayout>
   );
