+++ app/[locale]/community/news/helper/EditNewsPageContent.tsx
