+++ app/[locale]/community/news/helper/NewsPageContent.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import NewsRow from '@/app/[locale]/community/news/NewsRow';
+import NewsRow from '@/app/[locale]/community/news/helper/NewsRow';
 
 import LoginVisible from '@/components/common/LoginVisible';
 import Pagination from '@/components/common/Pagination';
@@ -28,6 +28,7 @@ export default function NewsPageContent({
 }) {
   const { page, keyword, tags, setSearchParams } = useCustomSearchParams();
   const { screenType } = useResponsive();
+  const pageLimit = screenType === 'desktop' ? 10 : 5;
 
   const setCurrentPage = (pageNum: number) => {
     setSearchParams({ purpose: 'navigation', pageNum });
@@ -36,14 +37,15 @@ export default function NewsPageContent({
   return (
     <PageLayout titleType="big">
       <SearchBox
-        key={tags + ''}
         tags={NEWS_TAGS}
         initTags={tags}
         initKeyword={keyword ?? ''}
         setSearchParams={setSearchParams}
       />
       <div className="mb-8 mt-10 flex flex-col gap-5 sm:mx-16">
-        {searchList.length > 0 ? (
+        {searchList.length === 0 ? (
+          <NoSearchResult />
+        ) : (
           searchList.map((post) => (
             <NewsRow
               key={post.id}
@@ -55,20 +57,21 @@ export default function NewsPageContent({
               imageURL={post.imageURL}
             />
           ))
-        ) : (
-          <p className="mx-2.5 mb-8 mt-6">검색 결과가 존재하지 않습니다.</p>
         )}
       </div>
       <Pagination
         totalPostsCount={total}
         postsCountPerPage={POST_LIMIT}
-        pageLimit={screenType == 'desktop' ? 10 : 5}
+        pageLimit={pageLimit}
         currentPage={page}
         setCurrentPage={setCurrentPage}
       />
+
       <LoginVisible staff>
         <AdminFeatures />
       </LoginVisible>
     </PageLayout>
   );
 }
+
+const NoSearchResult = () => <p className="mx-2.5 mb-8 mt-6">검색 결과가 존재하지 않습니다.</p>;
