+++ app/[locale]/community/news/helper/NewsRow.tsx
