+++ app/[locale]/search/SearchPageContent.tsx
@@ -4,7 +4,7 @@ import { useTranslations } from 'next-intl';
 
 import { Link } from '@/navigation';
 
-import NewsRow from '@/app/[locale]/community/news/NewsRow';
+import NewsRow from '@/app/[locale]/community/news/helper/NewsRow';
 import SeminarRow from '@/app/[locale]/community/seminar/SeminarRow';
 import NoticeRow from '@/app/[locale]/search/NoticeRow';
